Update and rename Minimum Sum Partition.py to Minimum_Sum_Partition.py
